move SolvEig function from main module to torsion module
